add some dto

+++ src/product/product.controller.ts
@@ -1,25 +1,38 @@
 import {
+  Body,
   Controller,
   Delete,
   Get,
   NotFoundException,
   Param,
+  Post,
 } from '@nestjs/common';
 import { Product as ProductModel } from '@prisma/client';
 import { ProductService } from './product.service';
 import { ProductModule } from './product.module';
+import { CreateProductDto } from './dto/create-product.dto';
 
 @Controller('product')
 export class ProductController {
   constructor(private readonly productService: ProductService) {}
 
-  @Get('/:id')
-  async getProductById(@Param('id') id: string): Promise<ProductModule> {
-    return this.productService.product({ id: Number(id) });
+  @Post('add')
+  async createProduct(@Body() body: CreateProductDto): Promise<ProductModel> {
+    return this.productService.createProduct(body);
   }
 
-  @Get('feed')
-  async getPublishedProducts(): Promise<ProductModel[]> {
+  @Get('id/:id')
+  async getProductById(@Param('id') id: string): Promise<ProductModule> {
+    const product = await this.productService.product({ id: Number(id) });
+    if (!product) {
+      throw new NotFoundException('Product not Found');
+    }
+
+    return product;
+  }
+
+  @Get('new')
+  async getisNewProducts(): Promise<ProductModel[]> {
     return this.productService.products({
       where: { isNew: true },
     });
@@ -51,6 +64,11 @@ export class ProductController {
 
   @Delete('/:id')
   async deleteProduct(@Param('id') id: string): Promise<ProductModel> {
+    const product = await this.productService.product({ id: Number(id) });
+    if (!product) {
+      throw new NotFoundException('Product not Found');
+    }
+
     return this.productService.deleteProduct({ id: Number(id) });
   }
 }
